fix(ci): split home summary contracts and drawer source
